+++ src/main/java/co/invest72/investment/domain/investment/CompoundFixedInstallmentSaving.java
@@ -165,7 +165,7 @@ public class CompoundFixedInstallmentSaving implements Investment {
 	}
 
 	@Override
-	public double getTaxRate() {
+	public BigDecimal getTaxRate() {
 		return taxable.getTaxRate();
 	}
 }
